Draft Code for Mespotamian Religion
Not sure how useful this'll be, but hopefully it's something and not too mismanaged. Put the features I could find code for in. Feel free to take in or out what you want. Figured intermarriage would work on the basis of Zoroastians, Egyptians, and Hellenics all being contemporaries and Tengri being a generic steppe religion as much as the Mongol-Turkic religion in particular, and the Mesoptoamians had a lot of interaction and intermixing with steppe nomads migrating into their land.